Cracked the stats problem.
We store stats in a format that's useful for display, and then we generate a hierarchical structure based on those settings, which we pass into all computed stats. The real special thing here is that computed stats are stored as getters inside that generated context, meaning we trade some small amount of performance for not having to calculate the dependency graph of said computed values.

Near as I can tell, it all just works. *Exhales*