converted server.
We do not babelify the server code anymore, updated all code to be ES2015 compliant, switched back to rethinkdb, and did a massive code cleanup of everything on the server. This brings the project inline with other recent projects.

Note: The UI is completely broken, pending a rewrite in Vue 2 and Vue Material.